backend/requirements.txt
# FastAPI & Server
fastapi==0.121.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
python-multipart

# Database - ASYNC (psycopg is pure Python, no C compiler needed)
sqlalchemy[asyncio]==2.0.25
psycopg[binary]==3.2.1
alembic==1.13.1

# Data processing
pandas
numpy

# HTTP - ASYNC
aiohttp==3.9.1

# Request validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.25.2
